requirements.txt
psycopg2
python-dotenv
pysqlite3
PIL